server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
motor
pydantic[email]
websockets
python-dotenv
